remove old table and rename current table component to SearchResultsTable
